fix(ci): first-party line limit — split five oversized test modules under 1500 lines
